+++ example/lib/main.dart
@@ -28,7 +28,6 @@ import 'package:z_components/components/z-expendable-item-tile.dart';
 void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
-
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -54,7 +53,7 @@ class MyApp extends StatelessWidget {
           onTap: (index) => index = index,
           currentIndex: 0,
           iconSize: 16,
-          inactiveColor: Colors.red,
+          inactiveColor: Colors.black,
           activeColor: Colors.black,
           selectedItemColor: Colors.white,
           unselectedItemColor: Colors.black45,
@@ -123,12 +122,13 @@ class _ComponentExemploClasseState extends State<ComponentExemploClasse> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        backgroundColor: const Color(0xffEFEFF4),
+        backgroundColor: const Color(0xff000000),
         floatingActionButton: ZFloatButton(
           onPressed: () {},
         ),
         appBar: ZNavigationBar(
           leading: new Icon(Icons.print),
+          middle: new Container(child: new Text('teste'),),
           trailing: new GestureDetector(
             onTap: () {
               Navigator.push(
@@ -322,7 +322,8 @@ class _ComponentExemploClasseState extends State<ComponentExemploClasse> {
                 Navigator.push(
                   context,
                   MaterialPageRoute(
-                      builder: (context) => ZCadastroUsuario(tituloAppBar: new Text("CADASTRO USUARIO"),
+                      builder: (context) => ZCadastroUsuario(
+                            tituloAppBar: new Text("CADASTRO USUARIO"),
                             onTapVoltar: () {
                               Navigator.of(context).pop();
                             },
@@ -348,7 +349,7 @@ class _ComponentExemploClasseState extends State<ComponentExemploClasse> {
                 );
               },
             ),
-            ZButton(
+            new ZButton(
               text: "PERFIL",
               onPressed: () {
                 Navigator.push(
@@ -377,6 +378,10 @@ class _ComponentExemploClasseState extends State<ComponentExemploClasse> {
                           )),
                 );
               },
+            ),
+            new ZButton(
+              zButtonType: ZButtonType.isCadastro,
+              onPressed: (){},
             ),
           ]),
         ]));